removed dependency form template

+++ other/sleep.hpp
@@ -1,1 +1,1 @@
-void sleep(ll t){ this_thread::sleep_for(chrono::milliseconds(t)); }
+void sleep(long long t){ this_thread::sleep_for(chrono::milliseconds(t)); }
